+++ lib/src/decoder/exception/invalid_escape_sequence.dart
@@ -1,5 +1,6 @@
 import 'package:meta/meta.dart';
 
+import '../../ast/value/string/escape.dart';
 import '../../exception.dart';
 
 /// An exception which is thrown when the parser encounters an unspecified
@@ -26,17 +27,48 @@ class TomlInvalidEscapeSequenceException extends TomlException {
   /// The invalid escape sequence.
   final String escapeSequence;
 
+  /// A human readable description of the reason why the escape sequence is
+  /// invalid.
+  final String reason;
+
   /// Creates a new exception for the given invalid [escapeSequence].
-  TomlInvalidEscapeSequenceException(this.escapeSequence);
+  TomlInvalidEscapeSequenceException._({
+    required this.escapeSequence,
+    required this.reason,
+  });
+
+  /// Creates a new exception for a non-Unicode [escapeSequence] that is not
+  /// one of [TomlEscapedChar.escapableChars].
+  factory TomlInvalidEscapeSequenceException.unspecified(
+    String escapeSequence,
+  ) {
+    var allowedEscapeSequences = TomlEscapedChar.escapableChars.keys
+        .map((escapableChar) => "\\$escapableChar")
+        .join(', ');
+    return TomlInvalidEscapeSequenceException._(
+      escapeSequence: escapeSequence,
+      reason: 'Must be one of $allowedEscapeSequences',
+    );
+  }
+
+  /// Creates a new exception for a Unicode [escapeSequence] that is not a
+  /// Unicode scalar value.
+  factory TomlInvalidEscapeSequenceException.nonScalar(String escapeSequence) =>
+      TomlInvalidEscapeSequenceException._(
+        escapeSequence: escapeSequence,
+        reason: 'Not a Unicode scalar value',
+      );
 
   @override
   bool operator ==(Object other) =>
       other is TomlInvalidEscapeSequenceException &&
-      other.escapeSequence == escapeSequence;
+      other.escapeSequence == escapeSequence &&
+      other.reason == reason;
 
   @override
-  int get hashCode => escapeSequence.hashCode;
+  int get hashCode => escapeSequence.hashCode ^ reason.hashCode;
 
   @override
-  String get message => 'The escape sequence "$escapeSequence" is invalid!';
+  String get message =>
+      'The escape sequence "$escapeSequence" is invalid: $reason';
 }