+++ lib/src/encoder.dart
@@ -1,5 +1,6 @@
 export 'encoder/ast_builder.dart';
 export 'encoder/encodable.dart';
+export 'encoder/exception/impossible_escape_sequence.dart';
 export 'encoder/exception/unknown_key_type.dart';
 export 'encoder/exception/unknown_value_type.dart';
 export 'encoder/pretty_printer.dart';